+++ titus-server-runtime/src/main/java/com/netflix/titus/runtime/connector/titusmaster/ConfigurationLeaderResolverComponent.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.titus.runtime.connector.titusmaster;
+
+import com.netflix.titus.common.runtime.TitusRuntime;
+import io.grpc.Channel;
+import io.grpc.netty.shaded.io.grpc.netty.NettyChannelBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ConfigurationLeaderResolverComponent {
+
+    @Bean
+    public LeaderResolver getLeaderResolver(TitusMasterClientConfiguration configuration) {
+        return new ConfigurationLeaderResolver(configuration);
+    }
+
+    @Bean(name = TitusMasterConnectorComponent.TITUS_MASTER_CHANNEL)
+    public Channel getManagedChannel(TitusMasterClientConfiguration configuration, LeaderResolver leaderResolver, TitusRuntime titusRuntime) {
+        return NettyChannelBuilder
+                .forTarget("leader://titusmaster")
+                .nameResolverFactory(new LeaderNameResolverFactory(leaderResolver, configuration.getMasterGrpcPort(), titusRuntime))
+                .usePlaintext(true)
+                .maxHeaderListSize(65536)
+                .build();
+    }
+}
